Fixing type of response time fields

+++ metricbeat/module/kibana/stats/data.go
@@ -54,10 +54,10 @@ var (
 		"request": RequestsDict,
 		"response_time": c.Dict("response_times", s.Schema{
 			"avg": s.Object{
-				"ms": c.Float("avg_ms", s.Optional),
+				"ms": c.Int("avg_ms", s.Optional),
 			},
 			"max": s.Object{
-				"ms": c.Float("max_ms"),
+				"ms": c.Int("max_ms"),
 			},
 		}),
 		"kibana": KibanaDict,
